motor_driver implemented for wheel and steering

+++ include/quadrature_encoder.hpp
@@ -58,7 +58,7 @@ public:
 
 private:
     PIO pio_instance;
-    int8_t sm;
+    uint sm;
 
     float ppr;
     float gear_ratio;
